Move some independent logic to `SetupBanner`.
